Replaced yacc generated parser with hand written one. Lexer still uses JFlex.
I was looking at the actual BNF grammar for the parser and it seemed so simple that a lot of overhead could be saved by using a hand-written one. Had a go and it worked. It still uses the JFlex generated lexer. Performance-wise it's much faster and uses less memory. This also removes the binary yacc executable complexities from the build.