Steps in architectural refactoring for SM3. Mostly work around: - Decoupling from Servlet environment. - Allow for non RequestDispatcher based decorators (e.g. direct FreeMarker rendering) - Clarifying domain concepts and API. - Simplifying core Filter.
All the current SM2 stuff will work as is at the moment. Much of the new SM3 stuff is built on top of this, with the intention of killing it off (except for publicly used APIs which will remain with adaptors for backwards compatability).

Note: SM3 stuff lives in new com.opensymphony.sitemesh package (not com.opensymphony.module.sitemesh).